safer teardown

+++ src/pytest_workflow/plugin.py
@@ -193,17 +193,19 @@ class WorkflowTestsCollector(pytest.Collector):
             # When we want to keep the workflow directory, write the logs to
             # the workflow directory.
             # TerminalReporter is used because print does not work.
-            log_err = self.workflow.stderr_to_file()
-            log_out = self.workflow.stdout_to_file()
-            # Print statements do not work here.
-            self.terminal_reporter.write_line(
-                "'{0}' stdout saved in: {1}".format(
-                    self.name, str(log_out)))
-            self.terminal_reporter.write_line(
-                "'{0}' stderr saved in: {1}".format(
-                    self.name, str(log_err)))
+            if self.workflow is not None:
+                log_err = self.workflow.stderr_to_file()
+                log_out = self.workflow.stdout_to_file()
+                # Print statements do not work here.
+                self.terminal_reporter.write_line(
+                    "'{0}' stdout saved in: {1}".format(
+                        self.name, str(log_out)))
+                self.terminal_reporter.write_line(
+                    "'{0}' stderr saved in: {1}".format(
+                        self.name, str(log_err)))
         else:
-            shutil.rmtree(str(self.tempdir))
+            if self.tempdir is not None:
+                shutil.rmtree(str(self.tempdir))
 
 
 class ExitCodeTest(pytest.Item):
